+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changes are grouped as follows
 
 ### Fixed
 - Fixed Templates API
+- `FunctionsAPI.list()` and `FunctionSchedulesAPI.list()` with argument `limit` equal to `None`, `float(inf)` or `-1` now returns all resources. Previously, the default limit of the API was used (100). 
 
 ## [0.50.0]